Add non-functional toggle mutation
Mutations to the backend API will be performed via Axios. However, the
newly added toggle mutation is not functional yet as I have not modified
the mutation to follow the JSON API format. Honestly, the JSON API
format feels like a bit of an overkill now, and I might decide to remove
it soon, as it adds quite a bit of complexity to the frontend code.