Add sorting and filtering capabilities
Sorting is performed using `lodash-es`, in case we want to add more
advanced behaviour in the future. The sorting and filtering states are
stored using React context and state. This is because we only have a
small number of global state items (2, maybe 3 if we at
ascending/descending order), and it doesn't make sense to introduce a
more complex system like Redux.