Frontend hooks: modify to use `/tasks/:id` route
This commit accompanies the previous commit in defaulting to using the
`/tasks` route in `useTasks` to prepopulate data for each individual
`useTask`, but each individual `useTask` hook can still fetch updated
data from `/tasks/:id`, which is handled by the excellent react-query.

The cache modification code for optimistic updates is also changed to
account for the new queries in `useTask`.